requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
python-multipart
pandas
openpyxl
python-dateutil
numpy
pydantic
pytest
